Reaching the root test of flakey resize
