+++ internal/driver/glfw/glfw_test.go
@@ -3,12 +3,32 @@
 package glfw
 
 import (
+	"runtime"
 	"testing"
 	"time"
 
 	"fyne.io/fyne/v2"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+func assertCanvasSize(t *testing.T, w *safeWindow, size fyne.Size) {
+	if runtime.GOOS == "linux" {
+		// TODO: find the root cause for these problems and solve them without additional repaint
+		// fixes issues where the window does not have the correct size
+		waitForCanvasSize(t, w, size, false)
+	}
+	assert.Equal(t, size, w.Canvas().Size())
+}
+
+func ensureCanvasSize(t *testing.T, w *safeWindow, size fyne.Size) {
+	if runtime.GOOS == "linux" || runtime.GOOS == "darwin" {
+		// TODO: find the root cause for these problems and solve them without additional repaint
+		// fixes issues where the window does not have the correct size
+		waitForCanvasSize(t, w, size, true)
+	}
+	require.Equal(t, size, w.Canvas().Size())
+}
 
 func repaintWindow(w *safeWindow) {
 	runOnMain(func() {
